requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
aiosmtpd
beautifulsoup4
pyyaml